pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Analises web pages"
authors = ["RZen <bartellbe@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
gunicorn = "^20.1.0"
validators = "^0.20.0"
psycopg2-binary = "^2.9.6"
requests = "^2.29.0"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
